Add files via upload
What Is a Backorder?
A backorder is an order for a good or service that cannot be filled at the current time due to a lack of available supply. The item may not be held in the company's available inventory but could still be in production, or the company may need to still manufacture more of the product.
The backorder is an indication that demand for a company's product outweighs its supply. They may also be known as the company's backlog. Also, company cannot overstock every product in their inventory to avoid such situation.
Problems with Backorders
If a company consistently sees items in backorder, this could be taken as a signal that the company's operations are far too lean. It may also mean the company is losing out on business by not providing the products demanded by its customers. If a customer sees products on backorder—and notices this frequently—they may decide to cancel orders, forcing the company to issue refunds and readjust their books.